Add settings loader class and interface
Allows for testing settings
#715
