+++ trview.app/Settings/ISettingsLoader.cpp
@@ -1,0 +1,8 @@
+#include "ISettingsLoader.h"
+
+namespace trview
+{
+    ISettingsLoader::~ISettingsLoader()
+    {
+    }
+}
